+++ src/main/java/dinkplugin/notifiers/SpeedrunNotifier.java
@@ -24,7 +24,7 @@ public class SpeedrunNotifier extends BaseNotifier {
     @Override
     public boolean isEnabled() {
         // intentionally doesn't call super as WorldUtils.isIgnoredWorld includes speedrunning
-        return config.notifySpeedrun() && settingsManager.testUsername(client.getLocalPlayer().getName());
+        return config.notifySpeedrun() && settingsManager.isNamePermitted(client.getLocalPlayer().getName());
     }
 
     @Override